+++ src/belote/__init__.py
@@ -1,3 +1,3 @@
-__version__ = "4.9.4"
+__version__ = "4.9.5"
 
 __all__ = ["__version__"]